Moved all scripts to invoke commands
